requirements.txt
gradio
opencc-python-reimplemented
pandas